remove extra code

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: stiffiny <stiffiny@student.21-school.ru>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/25 10:16:37 by stiffiny          #+#    #+#             */
-/*   Updated: 2021/05/05 11:50:55 by stiffiny         ###   ########.fr       */
+/*   Updated: 2021/05/06 21:14:22 by stiffiny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_calloc(size_t nmemb, size_t size)
 {
-	unsigned int		arr_size;
-	unsigned int		cur;
+	size_t			arr_size;
+	size_t			cur;
 	void				*p;
 
 	arr_size = 0;
